requirements.txt
asgiref==3.2.7
astroid==2.0
autopep8==1.5.3
certifi==2020.6.20
colorama==0.4.3
cycler==0.10.0
Django==3.0.4
django-betterforms==1.2
django-ckeditor==5.9.0
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-crum==0.7.5
django-datetimepicker==3.14
django-discover-runner==1.0
django-filter==2.3.0
django-js-asset==1.2.2
django-model-utils==4.0.0
django-role-permissions==3.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
freeze==1.0.10
html5lib==1.0.1
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
matplotlib==3.3.1
mccabe==0.6.1
numpy==1.19.2
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pylint==2.5.3
pylint-django==2.1.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2019.3
reportlab==3.5.42
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
Unipath==1.1
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.4
